requirements.txt
# TensorFlow (Apple Silicon)
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0

# Core ML/Data Processing
numpy==1.26.3
scikit-learn==1.4.1.post1
pillow==10.2.0
click==8.1.7

# Web Server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
jinja2==3.1.3
python-dateutil==2.8.2

# Optional: Development Tools
pytest==8.0.0
black==24.1.1
flake8==7.0.0